Slices without reflect
This PR changes the assembly code to receive pointers to the first bytes
of the slices instead of the slices themselves. This allows the Go
binding to simply pass pointers to the first byte and be able to expose
functions that take either `[][32]byte` or `[]byte`.

The generic fallback does copy chunks and digests in the case of byte
slice arguments.